+++ RELEASE_NOTES.md
@@ -4,7 +4,11 @@
 
 ## Version 1.0.3
 
-Implemented sharpness level reduction.
+- Implemented sharpness level reduction.
+- Implemented innate razor sharp.
+    - DB has ~66% chance to not lose sharpness.
+    - IG has ~33% chance to not lose sharpness.
+    - No other weapon is known to have innate razor sharp.
 
 ## Version 1.0.2